+++ example/example.html
@@ -8,6 +8,7 @@
   <select id="example-select">
     <option value="simple">simple</option>
     <option value="shape">shape</option>
+    <option value="shader">shader</option>
   </select>
   <div id="root"></div>
   <script src="./bundle.js"></script>
