+++ shaders/color.frag
@@ -1,0 +1,7 @@
+precision highp float;
+
+uniform vec4 color;
+
+void main(void) {
+  gl_FragColor = color;
+}
