+++ src/Shape.ts
@@ -1,7 +1,6 @@
 import {Vec2, Rect, Transform} from "paintvec"
 import {Context} from './Context'
 import {ObjectMap} from "./utils"
-import {Shader, UniformValue} from "./Shader"
 import {Color} from "./Color"
 import {Texture} from "./Texture"
