+++ src/index.ts
@@ -1,7 +1,7 @@
 export {Color} from "./Color"
 export {Context} from "./Context"
 export {DrawTarget, CanvasDrawTarget, TextureDrawTarget} from "./DrawTarget"
-export {Shader, ColorShader, TextureShader} from "./Shader"
+export {Shader, colorShader, textureShader} from "./Shader"
 export {PixelType, PixelFormat, TextureFilter, Texture} from "./Texture"
 export {Shape, QuadShape, RectShape} from "./Shape"
 export {BlendMode, Model, ShapeModelMode, ShapeModel, TextureModel} from "./Model"
